fix: websocket close frame
Fix sending a close frame on websocket disconnect, bug was found due to
dankchat's websocket library being very strict with closing frames.
Upstream bug report
https://github.com/snapview/tungstenite-rs/issues/405

Tungstenite does not flush the frame buffer on close, so when it echo's
back a close to the client it does not send that close unless we
explicitly call flush on the buffer.