security(api): gql get vulnerabilty
There is no reason to use GET requests in a web-browser, and it's likely a
CSRF attempt. We do not allow this because someone can redirect a user to
this endpoint, and they won't know or could embed it as an iframe or image
tag or something. The user would not know and it would execute graphql
queries on their behalf. Therefore we deny any GET request with a cookie,
origin, or referrer header.