fix: subscription refresh after gift
Move the subscription refresh to a seperate task after a sub has been
processed.

We schedule a task to run eventually to refresh subscriptions because
the updates may not reflect immediately in the database. This is a
bit-of-a hack however not a big deal since even if this does not run
the subscriptions will be refreshed eventually by the cron job.